requirements.txt
python == 3.9.19
smplx == 0.1.28
matplotlib == 3.9.1
scipy == 1.13.1
rerun-sdk == 0.17.0
trimesh == 4.4.2
mesh_to_sdf == 0.0.15
pyglet < 1.5.28
PyBullet == 3.2.6
mujoco == 3.1.6
mujoco-python-viewer == 0.1.4
google == 3.0.0
mediapy == 1.2.2
tqdm == 4.66.4
cma == 3.4.0
regex == 2024.5.15